requirements.txt
openai-whisper>=20231117
torch>=2.0.0
scikit-learn>=1.3.0
streamlit>=1.28.0
numpy>=1.24.0
hypothesis>=6.82.0
pytest>=7.4.0
